raspberry_pi/requirements.txt
python-dotenv
pymongo
Adafruit_DHT; sys_platform == "linux"